TMS delete button, password updates UI and API calls added

+++ src/app/components/users-view/projects-view/projects-view.component.ts
@@ -23,6 +23,8 @@ export class ProjectsViewComponent implements OnInit {
   timeZone: string;
   new_password: string;
   error_message: string;
+  error: boolean;
+  loading: boolean;
   public defaultTimeZone: string;
   updating_tms: boolean;
   // isLoggedInStatus = false;
@@ -36,6 +38,7 @@ export class ProjectsViewComponent implements OnInit {
     this.defaultTimeZone = 'GMT +7';
     this.showAdvancedSetting = false;
     this.updating_tms = false;
+    this.error = false;
     this.etabotAPI.get_real_projects();
     etabotAPI.projects.subscribe(data => this.setProjects(data));
     etabotAPI.projects.subscribe(change => this.setGotProjects());
@@ -64,6 +67,7 @@ export class ProjectsViewComponent implements OnInit {
     .subscribe(
       success => {
         this.updating_tms = false;
+        this.error = false;
       },
       error => {
         this.error_message = error;
@@ -82,6 +86,7 @@ Please enter another one or edit your existing one in projects screen.';
             }
         }
         console.log(error);
+        this.error = true;
         this.updating_tms = false;
       }
     );
@@ -89,33 +94,37 @@ Please enter another one or edit your existing one in projects screen.';
 
   delete_tms(tms_id) {
       // console.log('updating tms id ' + tms_id + ' with new password: ' + this.new_password);
-    this.updating_tms = true;
-    this.jiraService.delete_tms(tms_id)
-    .subscribe(
-      success => {
-        this.updating_tms = false;
-        this.etabotAPI.get_real_projects();
-      },
-      error => {
-        this.error_message = error;
-        if (String(error.status) === '400') {
-            if (error._body.includes('Unauthorized (401)')) {
-                this.error_message = 'Wrong combination of username/email and password. Please correct and try again.';
-            } else {
-                if (error._body.includes('already exists for this user')) {
-                    this.error_message = 'This username and team name already exist in your account. \
-Please enter another one or edit your existing one in projects screen.';
+    if (confirm('Are you sure you want to remove this JIRA account and its projects from ETAbot? You can add it again later.')) {
+        this.updating_tms = true;
+        this.jiraService.delete_tms(tms_id)
+        .subscribe(
+          success => {
+            this.updating_tms = false;
+            this.error = false;
+            this.etabotAPI.get_real_projects();
+          },
+          error => {
+            this.error_message = error;
+            if (String(error.status) === '400') {
+                if (error._body.includes('Unauthorized (401)')) {
+                    this.error_message = 'Wrong combination of username/email and password. Please correct and try again.';
                 } else {
-                    this.error_message = 'Bad request (4xx) - please check\
-                        all inputs and try again. If the issue persists, please report the issue to \
-                        hello@etabot.ai';
+                    if (error._body.includes('already exists for this user')) {
+                        this.error_message = 'This username and team name already exist in your account. \
+    Please enter another one or edit your existing one in projects screen.';
+                    } else {
+                        this.error_message = 'Bad request (4xx) - please check\
+                            all inputs and try again. If the issue persists, please report the issue to \
+                            hello@etabot.ai';
+                    }
                 }
             }
-        }
-        console.log(error);
-        this.updating_tms = false;
-      }
-    );
+            console.log(error);
+            this.error = true;
+            this.updating_tms = false;
+          }
+        );
+    }
   }
